backend/requirements.txt
fastapi
uvicorn
anthropic
pydantic
pydantic-settings
python-dotenv
